fix: Carmack-review pass — real low-risk fixes
- Total pack voltage decoded signed (int32) like current, matching JK encoding.
- Web /toggle + /setparam: proper pack bounds (reject b<0 || b>=numBms) instead
  of an &1 mask that silently remapped out-of-range indices.
- Web portal shows a red 'disconnected — retrying' state instead of silently
  going stale when /api fails (device reboot/WiFi blip).
- Log ArduinoOTA errors to serial instead of swallowing them.
- index.html replay tool: escape parse-error text (XSS from pasted input).

Reviewed and intentionally NOT changed: setPut write-register formula (the
'off-by-2' finding is a false positive — byte-granular JK write addressing is
verified working on hardware: capacity + max-charge writes persist). Noted as
follow-ups: split main.cpp along its section seams, optional powCv/capCv->PSRAM,
dynamic cell count for >4S packs.

+++ esp32-bms-lvgl/src/main.cpp
@@ -286,7 +286,7 @@ static void bmsReadAddr(uint8_t addr, int idx) {
     auto U32 = [&](int o) { return (uint32_t)((uint32_t)p[o] << 24 | p[o + 1] << 16 | p[o + 2] << 8 | p[o + 3]); };
     Bms &b = bms[idx];
     for (int i = 0; i < NCELLS; i++) b.cell[i] = U16(i * 2) / 1000.0f;
-    b.v = U32(0x90) / 1000.0f;
+    b.v = (int32_t)U32(0x90) / 1000.0f;   // signed like current, matches the JK i32 encoding
     b.i = (int32_t)U32(0x98) / 1000.0f;
     b.tMos = (int16_t)U16(0x8A) / 10.0f;
     b.tp1 = (int16_t)U16(0x9C) / 10.0f;
